+++ ChangeLog.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Reduced import time by waiting to import heavier packages pyvisa and pandas
 - lb.notebook is no longer pulled in by default; importing it now injects wrappers around builtins.range and np.linspace 
 - `host_log.txt` is now in YAML
+- `CommandLineWrapper` is now `ShellBackend`
 
 ### Removed
 - FilenameDict and ConcurrentRunner, which have been deprecated for a while